requirements.txt
Flask
Flask-SQLAlchemy
pymysql
Werkzeug
pyJWT
flask_wtf
Flask-Caching
python-dotenv
requests
mysqlclient
retrying
python-decouple
Flask-cors
gunicorn
Flask-Compress
black
flask_restx